move command prompt into own file
